fix: update export path build component
