+++ package.json
@@ -18,10 +18,10 @@
     "build:unpkg": "cross-env NODE_ENV=production rollup --config build/rollup.config.js --format iife",
     "postbuild": "rimraf ./dist/types/dev ./dist/types/src/entry.d.ts"
   },
-  "main": "dist/vue-pinch-scroll-zoom.ssr.js",
-  "browser": "dist/vue-pinch-scroll-zoom.esm.js",
-  "module": "dist/vue-pinch-scroll-zoom.esm.js",
-  "unpkg": "dist/vue-pinch-scroll-zoom.min.js",
+  "main": "dist/vue3-pinch-scroll-zoom.ssr.js",
+  "browser": "dist/vue3-pinch-scroll-zoom.esm.js",
+  "module": "dist/vue3-pinch-scroll-zoom.esm.js",
+  "unpkg": "dist/vue3-pinch-scroll-zoom.min.js",
   "types": "dist/types/src/entry.esm.d.ts",
   "files": [
     "dist/*",
