--- src/types.ts
@@ -1,17 +1,0 @@
-export interface PinchScrollZoomEmitData {
-    x: number;
-    y: number;
-    scale: number;
-    originX: number;
-    originY: number;
-    translateX: number,
-    translateY: number
-}
-
-export interface PinchScrollZoomSetData {
-  scale: number;
-  originX: number;
-  originY: number;
-  translateX: number,
-  translateY: number
-}
